SampleFormからSelectCategoryコンポーネントを独立させる
