+++ front/components/Category/SelectCategory.vue
@@ -1,0 +1,37 @@
+<template>
+  <v-card-subtitle class="ma-1">
+    <v-select
+      v-model="localCategoryId"
+      :items="categories"
+      item-text="name"
+      item-value="id"
+      label="カテゴリを選択"
+      style="max-width: 200px;"
+    ></v-select>
+  </v-card-subtitle>
+</template>
+
+<script>
+export default {
+  props: {
+    categoryId: {
+      type: Number,
+      required: true,
+    },
+    categories: {
+      type: Array,
+      required: true,
+    },
+  },
+  data() {
+    return {
+      localCategoryId: this.categoryId,
+    };
+  },
+  watch: {
+    localCategoryId: function (newCategoryId) {
+      this.$emit('updateCategory', newCategoryId);
+    },
+  },
+};
+</script>
