+++ front/pages/admin.vue
@@ -18,10 +18,12 @@
         <SampleForm
           :title="params.sample.title"
           :description="params.sample.description"
-          :category-id="params.sample.category_id"
-          :categories="categories"
           @updateTitle="params.sample.title = $event"
           @updateDescription="params.sample.description = $event"
+        />
+        <SelectCategory
+          :category-id="params.sample.category_id"
+          :categories="categories"
           @updateCategory="params.sample.category_id = $event"
         />
         <ChatLog
@@ -60,6 +62,7 @@
 import qs from 'qs';
 import PromptForm from '@/components/Sample/PromptForm.vue';
 import SampleForm from '@/components/Sample/SampleForm.vue';
+import SelectCategory from '@/components/Category/SelectCategory.vue';
 import ChatLog from '@/components/Sample/ChatLog.vue';
 import { handleFailure } from '@/plugins/error-handler';
 import checkAdminMixin from '@/plugins/check-admin-mixin';
@@ -68,6 +71,7 @@ export default {
   components: {
     PromptForm,
     SampleForm,
+    SelectCategory,
     ChatLog,
   },
   mixins: [checkAdminMixin],
